Move validator into resource middleware directory
